Update mcr.py
In some dataset mcr.py will create a C_temp with Nan values. Where C is unknown and an SVD guess is provided for ST. I didn't look into the problem futher, but this fix does solve the issue if present, and if not present the added code adds infinitesimally to the total run time.  Seems to be more of an issue in small datasets with very high S/N and somewhat small ST. 

If the Nan values are not removed, the script will crash at line 421